+++ NachosMedia/NachosMedia/Models/Genre.cs
@@ -8,11 +8,11 @@ namespace NachosMedia.Models
 {
     public class Genre
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
         [Required]
         public string Name { get; set; }
 
-        public byte? CategoryId { get; set; }
+        public int? CategoryId { get; set; }
     }
 }
